* 'master' of github.com:Coursemology/coursemology.org:
  multiple files upload for missions and training
  welcome page refactoring
  auto login app if Facebook is connected
  adding multiple file uploads control
  fixed resource loading error when customised theme incorrect
  correctes spelling mistake "staff" -> "stuff"
  allow stuff to edit students details
  enable edit of name for stuff management
  change date to date time for creating mission, training and announcement

+++ app/controllers/file_uploads_controller.rb
@@ -17,6 +17,7 @@ class FileUploadsController < ApplicationController
   end
 
   def create
+    puts "upload file",params
     if !current_user
       respond_to do |format|
         format.html { render text: "Unauthorized access!" }
@@ -28,30 +29,87 @@ class FileUploadsController < ApplicationController
       authorize! :upload_file, @course
     end
 
-    file_upload = FileUpload.create({
-      creator: current_user,
-      course: @course,
-      file: params[:files].first
-    })
+    if params[:file_upload].to_s.size > 0
+      file_upload = FileUpload.create({
+                                          creator: current_user,
+                                          file: params[:file_upload]
+                                      })
+      puts file_upload
+      puts "end here",@course
+    else
+      file_upload = FileUpload.create({
+                                          creator: current_user,
+                                          owner: @course,
+                                          file: params[:files].first
+                                      })
+      puts params[:files].first
+      puts "end here"
+    end
 
     if file_upload.save
-      puts file_upload.to_json
-      if file_upload.file_content_type == 'application/zip'
+      if file_upload.file_content_type == 'application/zip' && params[:_page_name] == "course_edit"
         course_theme = @course.course_themes.first_or_create
         theme_folder_url = course_theme.get_folder_url
         FileUtils.rm_rf(theme_folder_url)
         extract_theme_file(file_upload.file.path, theme_folder_url)
         resp = {
-          url: theme_folder_url
+            url: theme_folder_url
         }
       else
         resp = {
-          url: "#{request.scheme}://#{request.host_with_port}#{file_upload.file.url}"
+            url: "#{request.scheme}://#{request.host_with_port}#{file_upload.file.url}"
         }
       end
       respond_to do |format|
-        format.html { render json: resp }
+        if params[:_mode] == "MUL"
+          format.html {
+            render :json => [file_upload.to_jq_upload].to_json,
+                   :content_type => 'text/html',
+                   :layout => false
+          }
+          format.json { render json: {files: [file_upload.to_jq_upload]}, status: :created, location: file_upload }
+        else
+          format.html { render json: resp }
+        end
       end
     end
+  end
+
+  def index
+    puts "file upload index",params
+    owner = nil
+    if params[:training_id]
+      owner = Training.find(params[:training_id])
+    elsif params[:mission_id]
+      owner = Mission.find(params[:mission_id])
+    end
+
+    @uploads = owner ? owner.files : []
+
+    respond_to do |format|
+      format.html
+      format.json { render json: @uploads.map{|upload| upload.to_jq_upload } }
+    end
+  end
+  def new
+    @upload = FileUpload.new
+
+    respond_to do |format|
+      format.html # new.html.erb
+      format.json { render json: @upload }
+    end
+  end
+
+  def destroy
+    @upload = FileUpload.find(params[:id])
+    @upload.destroy
+
+    respond_to do |format|
+      #format.html { redirect_to uploads_url }
+      format.json { head :no_content }
+    end
+  end
+
+  def show
   end
 end
